add seo metadata
